# openapi/shamal-marafiq-v1.yaml
openapi: 3.1.0
info:
  title: Shamal FH2 External Viewer Middleware API
  version: 2.1.0
  description: |
    Shamal-controlled REST layer over DJI FlightHub 2 operations.
    External viewer integrators use **canonical** `/v1/viewer/*` paths with `X-Api-Key`.
    Legacy `/v1/marafiq/*` paths remain supported but are deprecated.
    Shamal owns FH2 credentials; external viewers never receive DJI access.
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
paths:
...
  /v1/marafiq/devices/{sn}:
    get:
      deprecated: true
      summary: "[Legacy] Device detail — use GET /v1/viewer/devices/{sn}"
      tags: [Devices, Legacy]
      parameters:
        - name: sn
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Device detail
  /v1/marafiq/devices/{sn}/telemetry/latest:
    get:
      deprecated: true
      summary: "[Legacy] Telemetry — use GET /v1/viewer/devices/{sn}/telemetry/latest"
      tags: [Devices, Legacy]
      parameters:
        - name: sn
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Telemetry snapshot
...
  /v1/marafiq/tasks/{id}:
    get:
      deprecated: true
      summary: "[Legacy] Task detail — use GET /v1/viewer/tasks/{id}"
      tags: [Tasks, Legacy]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Task detail
  /v1/marafiq/tasks/{id}/trajectory:
    get:
      deprecated: true
      summary: "[Legacy] Trajectory — use GET /v1/viewer/tasks/{id}/trajectory"
      tags: [Tasks, Legacy]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Trajectory
  /v1/marafiq/tasks/{id}/media:
    get:
      deprecated: true
      summary: "[Legacy] Task media — use GET /v1/viewer/tasks/{id}/media"
      tags: [Tasks, Legacy]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Media list
...
  /v1/marafiq/events/{id}/ack:
    post:
      deprecated: true
      summary: "[Legacy] Ack event — use POST /v1/viewer/events/{id}/ack"
      tags: [Events, Legacy]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Acknowledged
...